accessible/tests/mochitest/tree/test_tabbrowser.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>
<?xml-stylesheet href="chrome://browser/content/browser.css"
                 type="text/css"?>

<window xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        title="Accessible XUL tabbrowser hierarchy tests">

  <script type="application/javascript" 
          src="chrome://mochikit/content/MochiKit/packed.js" />
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js" />

  <script type="application/javascript"
          src="chrome://mochikit/content/a11y/accessible/common.js" />
  <script type="application/javascript"
          src="chrome://mochikit/content/a11y/accessible/role.js" />

  <script type="application/javascript">
  <![CDATA[
    ////////////////////////////////////////////////////////////////////////////
    // Test

    const Ci = Components.interfaces;

    // Hack to make xul:tabbrowser work.
    var XULBrowserWindow = {
      isBusy: false,
      setOverLink: function (link, b) {
      }
    };
    var gFindBar = {
      hidden: true
    };

    function doTest()
    {
      if (LINUX) {
        // XXX: bug 540529

        todo(false, "Failure on Linux.");
        SimpleTest.finish();
        return;
      }

      var tabBrowser = document.getElementById("tabbrowser");

      var progressListener =
      {
        onStateChange: function onStateChange(aWebProgress,
                                              aRequest,
                                              aStateFlags,
                                              aStatus)
       {
        if (aStateFlags & Ci.nsIWebProgressListener.STATE_STOP)
          testAccTree();
       }
      };

      tabBrowser.addProgressListener(progressListener,
                                     Ci.nsIWebProgress.NOTIFY_STATE_WINDOW);

      tabBrowser.loadTabs(["about:", "about:mozilla"], false, true);
    }

    function testAccTree()
    {
      var accTree = {
        role: ROLE_PANE,
        children: [
          {
            role: ROLE_TOOLTIP
          },
          {
            role: ROLE_MENUPOPUP
          },
          {
            role: ROLE_PAGETABLIST,
            children: [
              {
                role: ROLE_PAGETAB,
                children: [
                  {
                    role: ROLE_PUSHBUTTON
                  }
                ]
              },
              {
                role: ROLE_PAGETAB,
                children: [
                  {
                    role: ROLE_PUSHBUTTON
                  }
                ]
              },
              {
                role: ROLE_PUSHBUTTON
              },
              {
                role: ROLE_PUSHBUTTON
              }
            ]
          },
          {
            role: ROLE_PROPERTYPAGE
          },
          {
            role: ROLE_PROPERTYPAGE
          }
        ]
      };
      testAccessibleTree(getNode("tabbrowser").mTabBox, accTree);

      SimpleTest.finish()
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  ]]>
  </script>

  <hbox flex="1" style="overflow: auto;">
    <body xmlns="http://www.w3.org/1999/xhtml">
      <a target="_blank"
         href="https://bugzilla.mozilla.org/show_bug.cgi?id=540389"
         title=" WARNING: Bad accessible tree!: [tabbrowser tab] ">
        Mozilla Bug 540389
      </a><br/>
      <p id="display"></p>
      <div id="content" style="display: none">
      </div>
      <pre id="test">
      </pre>
    </body>

    <!-- Hack to make xul:tabbrowser work -->
    <menubar>
      <menu label="menu">
        <menupopup>
          <menuitem label="close window hook" id="menu_closeWindow"/>
          <menuitem label="close hook" id="menu_close"/>
        </menupopup>
      </menu>
    </menubar>

    <tabbrowser type="content-primary" flex="1" id="tabbrowser"/>
  </hbox>

</window>
